refactor(saga): remove unused saga classes and add stateless saga metadata parser test
- Remove SagaAnnotations.kt, SagaManager.kt, SagaMetadata.kt, SagaRepository.kt, and StatefulSaga.kt
- Add StatelessSagaMetadataParserTest.kt to test stateless saga metadata parsing
